Finalizado desenvolvimento do domínio deProfessor, iniciado o desenvolvimento do aluno.

+++ src/DTOs/FormacaoEnvioDTO.cs
@@ -10,5 +10,5 @@ public class FormacaoEnvioDTO
 
     public string AreaConhecimento { get; set; } = string.Empty;    
 
-    public decimal ValorPontuacao { get; set; }
+    public double ValorPontuacao { get; set; }
 }
